+++ src/clojure_lsp/shared.clj
@@ -11,6 +11,19 @@
 
 (def clojure-lsp-version (string/trim (slurp (io/resource "CLOJURE_LSP_VERSION"))))
 
+(defn assoc-some
+  "Assoc[iate] if the value is not nil. "
+  ([m k v]
+   (if (nil? v) m (assoc m k v)))
+  ([m k v & kvs]
+   (let [ret (assoc-some m k v)]
+     (if kvs
+       (if (next kvs)
+         (recur ret (first kvs) (second kvs) (nnext kvs))
+         (throw (IllegalArgumentException.
+                  "assoc-some expects even number of arguments after map/vector, found odd number")))
+       ret))))
+
 (def windows-os?
   (.contains (System/getProperty "os.name") "Windows"))